+++ src/fetcher/FetchResultTypes.ts
@@ -1,7 +1,7 @@
 import { TokenShortFragment } from '../graph-queries/uniswap-types';
 import {
   ReserveAuctionPartialFragment,
-} from '../graph-queries/zora-types';
+} from '../graph-queries/zora-graph-types';
 import {
   CurrentReserveBid,
   PastReserveBid,
@@ -12,6 +12,8 @@ export type MediaContentType =
   | { text: string; type: 'text'; mimeType: string };
 
 export type FetchGroupTypes = 'id' | 'creator' | 'owner';
+
+export type ZoraFetchQueryType = 'creator' | 'owner' | 'creator' | 'collection';
 
 type MetadataIsh = {
   mimeType: string;
@@ -50,6 +52,9 @@ export type NFTResultType = {
   contract: {
     address: string;
     knownContract?: KNOWN_CONTRACTS;
+    name?: string;
+    symbol?: string;
+    image?: string;
   },
   owner: ETHAddress;
   creator?: ETHAddress;